requirements.txt
matplotlib==3.0.3
opencv-python==4.1.1.26
scikit-image==0.15.0
scikit-learn==0.22.2.post1
tensorboard==2.0.2
py-sudoku==1.0.1
numpy==1.18.5